+++ rust-spice/tests/api/main.rs
@@ -29,6 +29,7 @@ mod frames;
 mod geometry;
 mod gf;
 mod orbits;
+mod planes;
 mod pool;
 mod rotations;
 mod spk;